Fix could not pull up equivalence class using projected target list" errors.
For example, from 'window' test:

regression=# explain (costs off)
select first_value(max(x)) over (), y
  from (select unique1 as x, ten+four as y from tenk1) ss
  group by y;
ERROR:  could not pull up equivalence class using projected target list (pathkeys.c:1314)

It was because even though cdbpullup_expr() would work with either a
target list (that is, a list of TargetEntrys), or a plain list of
expressions, it would behave slightly differently. With a plain list of
expressions, it would not match arbitrary expressions appearing in the
list.

I'm not sure if cdbpullup_exprs()'s behavior is sensible as it is. But
for now, let's just give it a target list.