Revert "Don't redistribute data for grouping, if the input is hashed on a constant."
This reverts commit ad2dd1af8733a40a4c50205f8850f1d7a520761d.

It was causing new failures, e.g. this:

select count(*) from tenk1 x where
  x.unique1 in (select a.f1 from int4_tbl a,float8_tbl b where a.f1=b.f1) and
  x.unique1 = 0 and
  x.unique1 in (select aa.f1 from int4_tbl aa,float8_tbl bb where aa.f1=bb.f1);
 count
-------
     0
     0
     1
(3 rows)

We know all the matching rows are on a single segment, but there's no
way to represent that knowledge, so that we would not run the aggregate
on the segments where we know are no matches.