Fix?  CURRENT OFF in the case that RLS is enforced
The assumption and test performed in getCurrentOf() for target relations does
not really match the current state of code.

Upstream commits <842faa714c0> and <b26e3d660df> altered the way that CURRENT
OF nodes work with respect to Row Level Security.

In greenplum the RangeTableEntries will have been expanded with the
subquery(ies) containing Row Level Security, if present causing
extractSimplyUpdatableRTEIndex to fail.

The inheritance cases are not visited at the moment and for this a fixme comment
is added.

The rowsecurity tests added on the abovementioned upstream commits regarding
getCurrentOf() are now passing.