Fix Orca aggregate plans with "intermediate" stages.
TODO: I'm a bit surprised that we only had a single test query for this.
I feel that we should have much more coverage for these complicated
multi-stage aggregates. Furthermoe, the test query that failed was
an upstream test, not even one that we had added in GPDB. That's also a bit
risky, because there's no guarantee that the test isn't modified in the
upstream, changing it so that it no longer covers the same codepath in
Orca.

This is the only test query that exercised this:

select ten, count(four), sum(DISTINCT four) from onek
group by ten order by ten;
INFO:  GPORCA failed to produce a plan, falling back to planner
DETAIL:  GPDB Expression type: GPDB_96_MERGE_FIXME: Intermediate aggregate stage not implemented not supported in DXL