Fix visibility check of in-progress XID against a distributed snapshot.
In PostgreSQL 9.5 and below, visibility checks for XIDs in
HeapTupleSatisfiesMVCC() function, follow this pattern:

1. TransactionIdIsInProgress(xid)
2. TransactionIdDidCommit(xid)
3. `XidInMVCCSnapshot(xid)

In 9.6 (commit 8a7d070181), this was changed to:

1. `XidInMVCCSnapshot(xid)
2. `TransactionIdDidCommit(xid)

In Greenplum, we've modified XidInMVCCSnapshot(), to check against the
distributed snapshot, if the XID belongs to a distributed transaction. As
part of that, XidInMVCCSnapshot() does this:

        /*
         * First, check if this committed transaction is a distributed committed
         * transaction and should be evaluated against the distributed snapshot
         * instead.
         */
        distributedSnapshotCommitted =
            DistributedSnapshotWithLocalMapping_CommittedTest(
                &snapshot->distribSnapshotWithLocalMapping,
                xid, false);

Note that the comment says "if this committed transaction ....". That
assumption used to be true; we used to only call XidInMVCCSnapshot() on
transactions that were already determined to be committed. But that
changed in 9.6, we now call XidInMVCCSnapshot() first, before checking if
it committed or not.

When DistributedSnapshotWithLocalMapping_CommittedTest() doesn't find the
XID in the distributed commit log, it concludes that it must have been a
local transaction, and marks it as committed in the backend local cache,
and also sets the HEAP_XMIN/XMAX_DISTRIBUTED_SNAPSHOT_IGNORE hint bit on
the tuple. That is wrong, now that the XID might belong to a transaction
that's still in-progress.

Fix by changing DistributedSnapshotWithLocalMapping_CommittedTest() to
not assume that the given XID was committed.

This showed up as occasional failures on the concourse pipeline like this:

--- /tmp/build/e18b2f02/gpdb_src/src/test/isolation2/expected/reindex/serializable_reindex_with_drop_index_ao.out    2019-10-30 20:20:27.831052228 +0000
+++ /tmp/build/e18b2f02/gpdb_src/src/test/isolation2/results/reindex/serializable_reindex_with_drop_index_ao.out    2019-10-30 20:20:27.831052228 +0000
@@ -48,7 +48,7 @@
 -- Remember index relfilenodes from master and segments before
 -- reindex.
 2: create temp table old_relfilenodes as (select gp_segment_id as dbid, relfilenode, oid, relname from gp_dist_random('pg_class') where relname like 'idx%_reindex_serialize_tab_ao' union all select gp_segment_id as dbid, relfilenode, oid, relname from pg_class where relname like 'idx%_reindex_serialize_tab_ao');
-CREATE 28
+CREATE 26

That should be fixed now.