Better fix for the flaky SEGV in 'starve_case' regression test.
The root cause is that the code in resqueue.c that sets up PROCLOCK
structs, like the upstream code in lock.c does, didn't get the memo about
the changes in lock.c in 9.6 that added the new groupLeader field. Fill
it properly in resqueue.c, too.

Also add a test case for it. With this isolation2 case, I could reproduce
the crash every time.