Fix conflicts in buffile.c.
I replaced it with the upstream version. Upstream got new mechanism to
share BufFiles across processes, which conflicted quite heavily with the
modifications we had done in GPDB. Many of the GPDB changes were also
related to sharing BufFiles across processes, so the plan is to change the
callers to use the new upstream SharedFileSet stuff for the sharing. Even
with that, we lost some babies with the bathwater: workfile manager stuff,
compression, and suspend/resume. Those are marked with FIXMEs.