Fix conflicts in nodeWindowAgg.c.
The GPDB added support for non-constant start/end offsets is almost
certainly broken. At a quick pass-through, I wasn't 100% sure which GPDB
added code was related to the VALUES preceding/following stuff that was
subsequently implemented in upstream, and thus should be replaced in the
merge, and which parts were related to the non-constant start/end offsets
feature. We'll find out when we get to run the regression tests...