Rename src/test/kerberos test suite, to avoid clashing with upstream.
PostgreSQL v11 added a kerberos test suite under src/test/kerberos. GPDB
has had a kerberos test suite of its own under the same path. That's
confusing, so rename the GPDB test suite.

We might want to get rid of the GPDB test suite altogether, as we merge
with PostgreSQL v11. But meanwhile, this allows us to keep them
side-by-side, which will avoid some confusion during the merge even if we
end up removing it in the end.