Fix crash from trying to pfree() invalid MinimalTuple.
TRCheckAndRemap() converted a HeapTuple into a MinimalTuple by playing
with the offsets. That gives you a valid pointer that can be accessed
like a MinimalTuple, but pfree()ing it doesn't work. Bite the bullet
and add a copy of TRRemapTuple() that contructs a MinimalTuple instead
of a HeapTuple.

Fixes crash in 'rowtypes' test.