Fix ALTER TABLE dispatching.
Previously, we handled all the AT_*Recurse variants in ATPrepCmd, because
QD changed the command types of some commands into the recurse variants
before dispatching. But I think it's simpler if we make a copy of the
whole ALTER TABLE command before modifying it, and dispatch the original.
I'm not sure if some if ATPrepCmd does some preparation that should be
dispatched, which would screw up this idea, but let's see how far we get
with this.