Resurrect HeapTuple->t_tableOid.
We had removed it in previous GPDB versions for performance reasons, but
we're putting it back now. I don't believe that the overhead is really
noticeable, at least not in recent versions.

NOTE: add a note in commit message about this. Or even better submit a
separate PR for this.