Revert "Fix inet case"
This reverts commit b13a4601f6bcfff9e25f0675807745c4136a7702.

Note that the test comment says "test index-only scans", but new plan is
not using an Index Only Scan. I'm not sure what's the best way to fix
that. Perhaps change the data in the table, or change the distribution key
so that the rows with same 'i' always come out in the same order. But I'll
revert that for now so that we don't forget about it.