Allow altering datatype of indexed column again.
This seems to work now. I think I've effectively reverted commit
fecd245aeb91 in the previous changes. I'm not sure how exactly it used to
not work, or how it works now.. But it seems all we need to do is to revert
the check that was put in place, and the test changes.