Fix demo cluster
Issue is that we had no longer loaded the 'internal.auto.conf' file. It
manifests itself during QD startup with following error:

```
FATAL, dbid (from -b option) is not specified or is invalid.  This value must
be >= 0, or >= -1 in utility mode
```

This happend after we got GPDB master commit a4b3caa3715c. RCA is that
we didn't have an entry in postgresqsl.config.sample to include
'special.conf' which is then later used as the replace value in initdb.c
to add 'internal.auto.conf'.