Remove gp_elog() function.
It was only used for one message in gprecoverseg, and it doesn't seem
important.

The second argument to the function didn't do anything, since the removal
of email and SNMP alerts in commit 65822b8019. And the NULL checks for the
arguments were pointless, because the function was marked as strict. But
rather than clean those up, let's just remove it altogether.

Reviewed-by: Asim R P <apraveen@pivotal.io>
Reviewed-by: Jimmy Yih <jyih@pivotal.io>