Accept changed query result.
The new result is in fact correct, and the previous one was wrong. Or at
least the new result seems more correct than the previous one. This is

I believe this changed with upstream commit 08f1e1f0a4, which was also
backported to 9.6. In any case, PostgreSQL produces the new result.