Fix fts incorrect parsing
When the mirror/standby is down and ftsprobes segment then assertion was
raised in ftsprobe.c:997

```
AssertImply(IsInSync, IsPrimaryAlive && IsMirrorAlive);
```

After looking at the bytes send over libpq we noticed that the fields
didn't align correctly. Need to investigate more as to why this is an
issue now on the merge branch. But for now commiting to fix crashes with
mirrors.

Co-authored-by: Ashwin Agrawal <aagrawal@pivotal.io>