Fix geometry test
Upstream commit a3d2844852dc6 improved the test coverage of geometric types.
Among them, a test case for division by zero was added which was largely based
on the order of the returned results being stable. In greenplum we can not make
such an unsafe assumption since we do not know which segment will return first.

Add a where clause that maintains the intention of the upstream commit.