Fix temp_relation test
In temp_relfilenode_primaries(), pg_ls_dir() is used as probe side of
the hash join and hash/build side is empty. It's empty because table
is dropped and hence lookup in pg_class is going to return zero rows.

Postgres 12 starts to inline CTE's hence pg_ls_dir() doesn't get
chance to fully complete and hence expected to not close the directory.

Hence, to validate if relfilenode is not leaked after dropping the
table, need to write to store the old relfilenode and not rely on
pg_class to fetch the value.

Co-authored-by: Jesse Zhang <sbjesse@gmail.com>