Accept some plan changes.
I did a manual EXPLAIN on one of these queries, and the cost of the bitmap
scan and index scan were the same. I guess the planner prefers the index
scan because it has a smaller startup cost. I'm not sure what what
changed, but the new plans with index scans seem just fine.