Detoast a slot before serializing it as MinimalTuple
We need to detoast a slot before serializing it into minimal tuple and send it
to master or other segments. The detoast should only happen on the same segment
which has the toast table, otherwise will get the below error:

```
ERROR:  missing chunk number 0 for toast value 60372 in pg_toast_60376 (tuptoaster.c:2294)
```

This patch first go through all the attributes of the slot to examine if any of
the attributes are external, if it has external attributes, we deform the slot
and detoast the external attributes and then form a minimal tuple.

Co-authored-by: Ashwin Agrawal <aagrawal@pivotal.io>