Accept expected output changes.
The meaningful difference here is that the planner is choosing a Hash Join
instead of Hash Semi Join in a couple of queries. That's because the
planner is smarter now, and can decude that the inner side of the join
produces at most one row, so a Hash Join is equivalent to a Hash Semi Join.