pg_rewind: Do not remove internal.auto.conf file
internal.auto.conf file is in the excluded file list for pg_rewind. Starting
from upstream commit a7eadaa ExcludeFiles semantics has changed. The new
semantic is that the ExcludeFiles are not be copied to and are deleted on the
target. For GPDB, we store gp_dbid in internal.auto.conf file, this file should
not be copied to the target but aslo should not be deleted from the target.
Therefore, excluding it from being removed from the target.

Co-authored-by: Ashwin Agrawal <aagrawal@pivotal.io>