WIP: appendonly select and insert using virtual tuple table slot
This code is not enabled yet, it needs grammar change to be enabled.
For insert, the table AM interface offers no way to maintain state
needed for appendonly insert operation.  We have worked around it by
defining a global object of type AppendOnlyInsertDesc.  A fixme is
added with more details.

It's possible that we are making redundant copies of the tuple during
insert or select.  This should be addressed as a followup.

Co-authored-by: Georgios Kokolatos <gkokolatos@pivotal.io>