WIP: allow gp_segment_id to be selected from appendoptimized table
Virutal tuple table slot does not have system attributes.  This patch
implements system attributes for appendoptimized tables using a new
tuple table slot type which is the same as virtual tuple table slot
except it has the getsysattr() interface defined differently.

Also added is a basic regression test for appendoptimized tables.
This is a temporary addition.  It can be removed once the existing
appendonly.source test is fixed in a follow up.