Remove appendonly and orientation from reloptions
There should be only one source of truth for the type of the relation and the
table access methods are prefered.

The legacy syntax has to continue working. For that end a this legacy layer has
been added in gram.y (currently turned off under a fixme) which takes care of
translation.

This commit removes a few hardcoded transformations and wax on layers.

Co-Authored-by: Asim R P <apraveen@pivotal.io>