Fix incorrect plan when a partitioned table was used in correlated subquery.
The correlation qual got lost, because apply_scanjoin_target_to_paths()
function would re-construct the AppendPath, but didn't re-construct
the Result to evaluate the "upperestrictinfo" that was added by
bring_to_outer_query() on the original AppendPath. Disable the
new upstream optimization in the unsafe cases.

This showed up as incorrect plans in a few queries in 'subselect_gp' test.