Remove GPDB tests for float8_regr_accum and functions backing aggregates.
The internal functions the implement regr_*() aggregates were changed in
v12, in commit e954a727f0, so that the values in the state array mean
different things than before. That broke the test. On a closer look, we
shouldn't need to be testing these functions in GPDB at all; they're
upstream functions, with no GPDB specific changes. We do have tests fo
the actual aggregates that these state functions are for, that should be
enough.

Per initial investigation and discussion on slack with Alex and Ashwin.