Revert "Change hot_standby default value to 'on'"
This reverts commit 34fc61673829f0cb63811fb63cc3c510fc2aadfc.

Reverting this commit as in GPDB we don't wish to enable hot_standby
yet. Also, this was causing
FailedAssertion(""!(PMAcceptingConnectionsStartTime) to be hit
sometimes for test `fts_recovery_in_progress` as with hot_standby
enabled the pmstate on mirror changes to PM_HOT_STANDBY instead of
PM_STARTUP. Due to this FTS probe requests which should get rejected
during mirror promotion goes through and hence hits
assertion. Ideally, we should look into how to avoid hitting this
assertion if hot_standby is enabled but that seems tomorrow problem.