Adjust expected output for error message changes.
All of these things were forbidden in the old GPDB partitioning
implementation, and are also forbidden in the new upstream implementation,
but the upstream error messages are different. Accept the upstream
messages. I'm not sure which ones are better, but either way, we don't want
to maintain a diff just for that.