Move MemTupleBinding initialization to when it is needed
This allows using the slot's tuple descriptor to create memtuple
binding.  Previously, the binding was created in beginscan, where slot
is not available, and the relation's tuple descriptor was being used.
Using relation's tuple descriptor is not correct because alter table
rewrite (in case of add column) changes the catalog before rewriting
the table.  The old relation's tuple descriptor contains newly addded
columns and therefore, it cannot be used to scan the old table.

Co-authored-by: Georgios Kokolatos <gkokolatos@pivotal.io>