Refactor stdRdRelopts and some of their methods for appendoptimized relations
Storage access methods used to be defined within stdRdRelopts. With the
introduction of tableam api, there should be only one place of truth. So the
access method options were removed from stdRdRelopts. As a consequence, the
appendoptimized related stdRdRelopts needed to be distinguished from the rest. A
new relopt kind was introduced for this purpose.

A side effect is that in heap_reloptions() do not carry any access method
information, so in some cases default_reloptions() need to be called directly.
This has been managed to be confined in DefineRelation() and greenplum specific
guc setting methods.

Also clean up some of the of wax on / wax off places of stdRdRelopts
transformations and validations.

As a side-effect, some of the pre-existing error messages will be altered.

Tests are not ammended in this commit to reflect this change.

Co-authored-by: Asim R P <apraveen@pivotal.io>