Implement DROP PARTITION FOR(<value>)
The main logic in this commit is to find target partition to operate
on based on value specified. We used the logic similar to what INSERT
uses. Actually, for leveraging same code made
get_partition_for_tuple() extern instead of static upstream.

Given now we have logic to find partition based on name and value,
this can be leveraged by other ALTER TABLE commands operating on
partitions as well.

Co-authored-by: Alexandra Wang <lewang@pivotal.io>