Reset file offest before switch to a new segment file
Previously, the FileRead() interface did not accept offset as a
parameter.  It used to be the caller's responsibility to seek the file
pointer to the right location before calling FileRead().  But now,
FileRead() takes care of performing seek, based on the offeset passed
as argument.  So let's reset it to 0 when we are starting to scan a
new segment file.

Co-authored-by: Georgios Kokolatos <gkokolatos@pivotal.io>