Use frozen insert for gp_fastsequence
If a new segment file needs to be created by an appendoptimized insert
transaction, a new entry in gp_fastsequence is needed.  This entry
must be inserted as a frozen tuple, so that it remains visible even if
the insert transaction aborts.  Fastsequence is used to hand out row
numbers and row numbers must be unique.

We had to introduce a public function in indexing.c for this purpose.
It uses CatalogIndexInsert, a static function defined in indexing.c.

Co-authored-by: Georgios Kokolatos <gkokolatos@pivotal.io>