Enable CTAS for appendoptimized relations
Leverage on the startup function provided to create the appendonly state needed
for insertions. Previously the appendoptimized state was held into the createas
state. Now it belongs in the private memory of the appendonlyam handler.

The appendoptimized handler does not seem to care about the bulk insert state
provided. The bulk insert state seems to be heavily heap related, yet it might
be beneficial for appendoptimized also. Investigate.

Co-authored-by: Asim R P <apraveen@pivotal.io>