Reinstate the behaviour of erroring out on duplicate ao storage options
We now strip away legacy AO storage options during parsing in favour of tableam
options. Previously an error was emmited if multiple instances of these options
were found. That behaviour got lost during stripping and the last occurance
would prevail. This commit reinstates the old behaviour.

A bit of caution has to be maintainted as to not grow too much the logic in that
part of the code as it lives in grammar. If it grows too much, it would be
better to be moved further down during processing.

	modified:   src/test/regress/output/appendonly.source