Lock relation only if it is AO/CO when creating blockdirectory
Previously, a relation was opened with AccessExclusiveLock and if it
was not found to be an AO/CO table, no block directory was created but
the lock was held until end of transaction.  This is now changed to
first obtain pg_class tuple of the relation from syscache and only if
it's AO/CO, it is opened with AccessExclusiveLock.

The reindex isolation2 test was getting stuck due to a deadlock
without this change.

Also added a fixme to see if a lock weaker than AccessExclusiveLock
can be used because concurrent select operations should be allowed
while blockdirectory is being created.