Relocate the fault point used by reindex isolation2 test
The previous location for the reindex_db fault point was rendered
unused by newly merged code.  Removed the unused function and moved
the fault to a new place where it is trigged according to how the test
expects.  The test used to get stuck until the
wait_for_fault_triggered() interface timed out.  Now completes in less
than 3 seconds and passes.

Also fixed a compiler warning in the same file.