Modify list_qsort() to use qsort_arg() instead of qsort()
To compare partition keys for mapping GPDB partitioning syntax to
postgres, we sort CreateStmt List. Need to have access to PartitionKey
for the same. Previously, list_qsort() didn't allow to specify any
argument to comparator function. Modifying the same to allow passing
arg to comparator.

Ideally, this change can be upstreamed but with no real usecase for
list_qsort(), would be hard to convince for change.