Fix assertion failure in 'bb_mpph' test.
create_minmax_plan() has side-effects, so it's not OK to call it twice.
Previously, cdbpathtoplan_create_sri_plan() would call create_plan(),
but bail out because it didn't produce a Result plan. Later, we called
creat_plan() again, but create_minmax_plan() had already set
root->minmax_paths, so you got an assertion failure.

Be more careful to only proceed with create_plan() if the path is a
GroupResultPath, so that it really will produce a trivial Result plan.