Create a memory context in appendonlyam_handler responsible for local state
During the implementation of the table am AO specific handler, a place to hold
the local state, e.g insert descriptor was needed. The choice was made to hold
that in the AO handler's private memory of the backend process. It is assumed
and heavily relied to at this moment, that only one concurrent DML statment can
be running per backend proccess. This assumption may well be false, so it is
needed to verify that the local state is NULL every time a DML is executed.

During a transaction abort, everything is taken care off except the local state.
A mechanism to reset the local state was needed. Enter
MemoryContextRegisterResetCallback(). Now we use this interface in order to
disassociate the local state from the process. This interface is also called in
the happy path and is a no-op.

Co-authored-by: Asim R P <apraveen@pivotal.io>