Implement multi insert for AO, a first pass
At this point the code is simply a loop over appendonly_tuple_insert() in order
to make the tests start passing. Levereages the exposed finish_bulk_insert
method which removes some switches over the relation's am in the upstream code.

Also removes ri_aoInsertDesc from ResultRelInfo. It is not wise to keep track of
the descriptors in one place for one sort of operations and another for the
rest. Since we do need to store the insert descriptor in the am's private memory
for insert commands, we favour it over ResultRelInfo.

Co-authored-by: Asim R P <apraveen@pivotal.io>