Reset the grammar/lexer hack to deal with CREATE TABLE PARTITON BY syntax.
In CREATE TABLE AS, the "tail_partition_magic" flag was not reset. That
caused a syntax error, if you had CREATE TABLE AS followed by another
statement with the PARTITION keyword, like a window function "SELECT sum()
OVER (PARTITION BY ...)", in the same statement. This was not reproducible
with psql, because psql sends the statements one-by-one, but gpcheckcat
was doing exactly that.