Enable RelOptions and tablespace for partitions
In both Greenplum master and upstream Postgres, the child partitions
always inherit the tablespace from the parent partition if the user
didn't explicitly specify one. If the user specified a different
tablespace for a partition, then the partition uses the user specified
tablespace.

RelOptions semantics are a little different between Greenplum master and
Postgres. Upstream does not allow specifying RelOptions for the parent
partitioned table, while Greenplum allows it. In Greenplum, the
partitions that are created from the CREATE TABLE <parent table>
statement inherit the same RelOptions, whereas the partitions that are
created from ALTER TABLE <parent table> ADD PARTITION don't inherit
RelOptions from the parent table. So the decision we made is to allow
the user to give a RelOptions if the Greenplum syntax is used, and
during execution, we ignore the RelOptions for the parent, but give the
child partitions that RelOptions.

Similar to parsing the regular CreateStmt, we used
greenplumLegacyAOoptions() to convert options to accessMethods for AO.

Also removed OptTabPartitionStorageAttr node from gram.y. We instead
used the upstream node OptWith and OptTableSpace.

Co-authored-by: Ashwin Agrawal <aagrawal@pivotal.io>