Handle WITH(tablename='...') for partitions
As part of this commit, we also added to make deep copy for options
list for every partition. It was messy and unsafe to have create stmt
for different partitions point to same options LIST.

As in old times, we also don't deal with multiple tablenames in
options list. It would ERROR out same way as before.

Co-authored-by: Alexandra Wang <lewang@pivotal.io>