Silence failures in 'window' test.
The output for these queries is not fully deterministic. It depends on
the order that tuplesort returns rows that have the same sort key value.

Plus one seemingly harmless plan change. The old plan looked OK, with
the 'depname' being eliminated from the Partitioning Key. But the new plan
is fine too. I'm not sure if we lost that optimization somewhere along the
line, but it doesn't seem important.