Fix sematic for Partition ID Name type
Legacy GPDB code stores partition name (which is essentially postfix
of full partition name) in catalog table pg_partition_rule. The full
partition name is geneated using
parenttablename_prt_<level>_<partname>. In new world we don't have
catalog table to store the partition name. Hence, currently we are
trying to use same semantic for partition name as CREATE TABLE
PARTITION or ADD PARTITION uses. To construct the name with same rules
and then try to find the partition corresponding to it.

Lets see if even after this poses the need for storing the partition
name aka postfix in catalog somewhere or we can deal without it.