Use tbm_bitmapword type when using TID bitmaps
The type tbm_bitmapword is Greenplum specific.  It represents 64-bit
bitmap words that are used for TID bitmaps.  PostgreSQL makes use of
bitmapword from bitmapset.h for TID bitmaps.  The bitmapword type used
to be 32-bits wide but in current merge iteration, it was widened to
64-bits if the underlying archiecture is 64-bit wide.

It is desirable to avoid this merge hazard by removing the special
type tbm_bitmapword and use bitmapword everywhere as in upstream.
That's not possible as yet, because for appendoptimized visimaps, the
bitmapword type is used to encode tuple visibility.  The visibility
information, including the bitmap words, is stored on disk, which
means bitmapword type cannot be widened to 64-bits without incurring
upgrade penalty.

For now, let's continue to use tbm_bitmapword as type for TID bitmaps
consistently.  In a follow up commit, bitmapset word size will be
changed to remain 32-bits wide to facilitate their usage in
appendoptimized visimaps.