Use the new GUC in tests to create appendoptimized tables by default
The gp_default_storage_options no longer accepts 'appendonly=true' and
'orientation=column'.  The new GUC default_table_access_method should
be used instead.  The access method name for row-oriented
appendoptimized tables is "appendoptimized" and for column-oriented
tables is "aoco".  This patch performs the necessary string
replacement.

The sql and answer file generation logic in pg_regress is also
changed accordingly.

There are many more tests to be changed, this is just a small step
forward.  I am working on appendoptimized update operation and in
order to test it along with insert and delete, I wanted to run some of
thses tests.