Accommodate multiple relations in appendoptimized DML state
The state maintained by appendoptimized access method needs to be able
to deal with multiple relations.  E.g. in case of partitioned tables,
a single insert statement on the root table may result in tuples being
routed to multiple partitions.

We have chosen a hash table, keyed by the relation's OID to maintain
the DML state (insert, update and delete descriptors).  To reduce the
cost of hash lookup per tuple, a reference to the most recently used
state is cached.

Update to distribution key (split update) is modeled as delete and
insert.  The access method interface sees this as an update operation.
We have to create the state necessary for performing insert as well as
delete in case of an update operation, because the access method
cannot tell the difference between a normal update and a split update.
Consequently, we had to expose the insert descriptor that is part of
update descriptor.

Co-authored-by: Georgios Kokolatos <gkokolatos@pivotal.io>