Enable appendoptimized and aoco access methods
Until now, appendoptimized and aoco access methods were disabled in
the grammar such that heap access method was silenty used.  Now that
change is removed.  The right access method will be used, according to
what was specified.

We have implemented appendoptimized access method to a sufficient
extent that we can deal with test failures and crashes, if any.  The
aoco access method is not implemented yet.  We expect to see the
corresponding tests fail with "not implemented" errors.

Co-authored-by: Georgios Kokolatos <gkokolatos@pivotal.io>