Refactor the internal state handling for dmls targeting ao relations
Admittedly this is not the final state of the code. What is now stored localy it
should be stored in an adecuate struct of the caller. However the tableam api
does not provide such an interface yet nor the callers provide a space for
carrying that information. In order to implement such a method properly, it is
desired to know where such an initialization would be needed, what structures we
have available to store that info and make certain that we deinitialize it
properly.

In the final step, any call outside of the tableam api that contains a check for
AppendOnly relation should disappear and any Appendonly specific function that
is leaked should be replaced by the properly added interface.

The api currenlty provides only a method for finalizing bulk insert. We are
piggy backing on that quite some, though it is not yet verified that we properly
deinitialize the state in all places.

Once most of the AO related tests are green, we will start implementing a new
table access method to initialize state and augment which ever other methods are
needed to take that state as an argument.

Co-authored-by: Asim R P <apraveen@pivotal.io>