Disable TID SCAN for appendoptimized relations.
There is not a viable way to perform TID SCAN's for appendoptimized relations.
The main reason for it is that a) a block directory might not exist for the
relation, b) even if it exists, the state needed to scan and determine
visibility is not available or cheaply created per tuple fetched. Look the
available interface tuple_fetch_row_version() for it's usage.

In greenplum 6, reloptinfo was holding a specific relstorage member to check for
it. Now we are storing the relam oid instead. An attempt to not access it
directly has been made.

Co-authored-by: Asim R P <apraveen@pivotal.io>