Use same function in 'commit_blocking_on_standby' and 'dtx_recovery_wait_lsn'
The two tests had used the same copy-pasted test function, but commit
26930987d changed the copy in 'commit_blocking_on_standby' test to have
a timeout, and also changed the return type. Changing the return type made
the CREATE OR REPLACE FUNCTION command in 'commit_blocking_on_standby' test
to fail, because you can't change the return type of a function with CREATE
OR REPLACE. Change the copy in 'commit_blocking_on_standby' to match the
one in 'dtx_recovery_wait_lsn'.

I'm not set up to run this test locally right now, so I did this blindly
and it might well be broken. I'll keep an eye at the pipeline.