Fix the length of tuple length when serializing a tuple for Motion.
The length that was transmitted was 4 bytes too large, because we used the
wrong variable here. Or alternatively, you could say that the receiver
understood the length wrong, but I chose to change the transmitting side.

This was surprisingly harmless. All tuples received by a Motion contained
4 bytes of garbage after the actual data, but that did little harm. If
the tuple was written to disk, the garbage bytes took space unnecessarily,
and that's what gave this away: The contrib/indexscan read_heap_index test
was failing, because not so many tuples fit on a single page and that
caused the CTIDs to differ from what was memorized in the expected output.